Updated MSVC project and cleaned up some more type conversion warnings. Fixed a path error for WIN32 systems with home directories residing on places other than the C:\.

+++ src/actions/boarder.hxx
@@ -32,8 +32,8 @@ class Boarder : public PinguAction
 {
 private:
   //double counter;
-  double x_pos;
-  double speed;
+  float x_pos;
+  float speed;
   StateSprite sprite;
 public:
   Boarder (Pingu* p);
